Add plugin support for rendering and backends
Remove rendering from the `Site` and `Page` and split it into
`Renderers` and `Backeneds` the default rendering chain is markdown -> jinja2.
The default backends is the file backend.

Allow Renderers and Backends to be loaded from `$HOME/.stone/`. These
can now be specified in the in the site.json config file.